fix: matches arrays when searching for nested fields
For a query like `{"a.b.c": 1}`, MongoDB matches a document where the 
field `"b"` is an array containing the object `{"c": 1}`.

This case is handled by this commit. We also need to investigate the
broader case where it should also match the following objects:

{"a": [{"b": {"c": 1}}]}
{"a": [{"b": [{"c": 1}]}]}

Those other cases are being tracked on issue #67
https://github.com/fcoury/oxide/issues/67